refactor: separate terminal window into component
